source/requirements.txt
requests==2.22.0
PyMySQL==0.9.3
chardet==3.0.4
psutil==5.6.3
DBUtils==1.3
kafka-python==1.4.6
redis==3.2.1
beautifulsoup4==4.6.0
PyExecJS==1.5.1
elasticsearch==7.0.2
